Remove fitserror
