+++ fitsio/src/macros.rs
@@ -1,7 +1,7 @@
 /// Macro to return a fits error if the fits file is not open in readwrite mode
 macro_rules! fits_check_readwrite {
     ($fitsfile: expr) => (
-        use $crate::fitserror::FitsError;
+        use $crate::errors::FitsError;
         if let Ok($crate::fitsfile::FileOpenMode::READONLY) = $fitsfile.open_mode() {
             return Err(FitsError {
                 status: 602,
